# `regex-lint` only reads files that will be checked